+++ tests/ZipEntryCmdlets.tests.ps1
@@ -110,12 +110,12 @@ Describe 'ZipEntry Cmdlets' {
         }
 
         It 'Can list zip file entries' {
-            $zip | Get-ZipEntry -EntryType Archive |
+            $zip | Get-ZipEntry -Type Archive |
                 Should -BeOfType ([PSCompression.ZipEntryFile])
         }
 
         It 'Can list zip directory entries' {
-            $zip | Get-ZipEntry -EntryType Directory |
+            $zip | Get-ZipEntry -Type Directory |
                 Should -BeOfType ([PSCompression.ZipEntryDirectory])
         }
 
@@ -133,39 +133,39 @@ Describe 'ZipEntry Cmdlets' {
 
     Context 'Get-ZipEntryContent' -Tag 'Get-ZipEntryContent' {
         It 'Can read content from zip file entries' {
-            $zip | Get-ZipEntry -EntryType Archive |
+            $zip | Get-ZipEntry -Type Archive |
                 Get-ZipEntryContent |
                 Should -BeOfType ([string])
         }
 
         It 'Can read bytes from zip file entries' {
-            $zip | Get-ZipEntry -EntryType Archive |
+            $zip | Get-ZipEntry -Type Archive |
                 Get-ZipEntryContent -AsByteStream |
                 Should -BeOfType ([byte])
         }
 
         It 'Can output a byte array when using the -Raw switch' {
-            $zip | Get-ZipEntry -EntryType Archive |
+            $zip | Get-ZipEntry -Type Archive |
                 Get-ZipEntryContent -AsByteStream -Raw |
                 Should -BeOfType ([byte[]])
         }
 
         It 'Should not attempt to read a directory entry' {
-            { $zip | Get-ZipEntry -EntryType Directory | Get-ZipEntry } |
+            { $zip | Get-ZipEntry -Type Directory | Get-ZipEntry } |
                 Should -Throw
         }
     }
 
     Context 'Remove-ZipEntry' -Tag 'Remove-ZipEntry' {
         It 'Can remove file entries' {
-            { $zip | Get-ZipEntry -EntryType Archive | Remove-ZipEntry } |
+            { $zip | Get-ZipEntry -Type Archive | Remove-ZipEntry } |
                 Should -Not -Throw
 
             $zip | Get-ZipEntry | Should -Not -BeOfType ([PSCompression.ZipEntryFile])
         }
 
         It 'Can remove directory entries' {
-            $entries = $zip | Get-ZipEntry -EntryType Directory
+            $entries = $zip | Get-ZipEntry -Type Directory
             { Remove-ZipEntry -InputObject $entries } | Should -Not -Throw
             $zip | Get-ZipEntry | Should -Not -BeOfType ([PSCompression.ZipEntryDirectory])
         }
@@ -239,43 +239,43 @@ Describe 'ZipEntry Cmdlets' {
         }
 
         It 'Can rename file entries using a delay-bind ScriptBlock' {
-            { $zip | Get-ZipEntry -EntryType Archive | Rename-ZipEntry -NewName { 'test' + $_.Name } } |
+            { $zip | Get-ZipEntry -Type Archive | Rename-ZipEntry -NewName { 'test' + $_.Name } } |
                 Should -Not -Throw
 
-            $zip | Get-ZipEntry -EntryType Archive |
+            $zip | Get-ZipEntry -Type Archive |
                 ForEach-Object Name |
                 Should -Match '^test'
         }
 
         It 'Produces output with -PassThru' {
-            $zip | Get-ZipEntry -EntryType Archive |
+            $zip | Get-ZipEntry -Type Archive |
                 Rename-ZipEntry -NewName { $_.Name -replace 'test' } -PassThru |
                 Should -BeOfType ([PSCompression.ZipEntryFile])
         }
 
         It 'Can rename directory entries and all its child entries' {
-            $dir = $zip | Get-ZipEntry -EntryType Directory -Include testfolder00/
+            $dir = $zip | Get-ZipEntry -Type Directory -Include testfolder00/
             $childs = $zip | Get-ZipEntry -Include testfolder00/*
             Rename-ZipEntry $dir -NewName myNewName
             $zip | Get-ZipEntry -Include myNewName/* | Should -HaveCount $childs.Count
         }
 
         It 'Should throw if an entry with the same Name already exists' {
-            { $zip | Get-ZipEntry -EntryType Directory -Include testfolder01/ |
+            { $zip | Get-ZipEntry -Type Directory -Include testfolder01/ |
                 Rename-ZipEntry -NewName testfolder02 } |
                 Should -Throw
 
-            { $zip | Get-ZipEntry -EntryType Archive -Include testfolder01/file00.txt |
+            { $zip | Get-ZipEntry -Type Archive -Include testfolder01/file00.txt |
                 Rename-ZipEntry -NewName file01.txt } |
                 Should -Throw
         }
 
         It 'Should throw if renaming an entry that no longer exists' {
-            $entry = $zip | Get-ZipEntry -EntryType Archive -Include testfolder01/file00.txt
+            $entry = $zip | Get-ZipEntry -Type Archive -Include testfolder01/file00.txt
             $entry | Remove-ZipEntry
             { $entry | Rename-ZipEntry -NewName test } | Should -Throw
 
-            $entry = $zip | Get-ZipEntry -EntryType Directory -Include testfolder01/
+            $entry = $zip | Get-ZipEntry -Type Directory -Include testfolder01/
             $entry | Remove-ZipEntry
             { $entry | Rename-ZipEntry -NewName test } | Should -Throw
         }
@@ -283,11 +283,11 @@ Describe 'ZipEntry Cmdlets' {
         It 'Should throw if using invalid FileName chars' {
             $invalidChars = [System.IO.Path]::GetInvalidFileNameChars()
 
-            { $zip | Get-ZipEntry -EntryType Archive |
+            { $zip | Get-ZipEntry -Type Archive |
                 Rename-ZipEntry -NewName { $_.Name + $invalidChars[0] } } |
                 Should -Throw
 
-            { $zip | Get-ZipEntry -EntryType Directory |
+            { $zip | Get-ZipEntry -Type Directory |
                 Rename-ZipEntry -NewName { $_.Name + $invalidChars[0] } } |
                 Should -Throw
         }