+++ CHANGELOG.md
@@ -4,11 +4,14 @@
         - Renamed Property `EntryRelativePath` to `RelativePath`.
         - Renamed Property `EntryType` to `Type`.
         - Renamed Method `RemoveEntry()` to `Remove()`.
+        - Added Property `CompressionRatio`.
     - `ZipEntryFile` Type:
         - Added Property `Extension`.
         - Added Property `BaseName`.
     - `ZipEntryDirectory` Type:
         - `.Name` Property now reflects the directory entries name instead of an empty string.
-    - Added Cmdlet `Rename-ZipEntry`.
+    - Added command `Rename-ZipEntry`.
     - `NormalizePath` Method:
         - Moved from `[PSCompression.ZipEntryExtensions]::NormalizePath` to `[PSCompression.Extensions.PathExtensions]::NormalizePath`.
+    - `Get-ZipEntry` command:
+        - Renamed Parameter `-EntryType` to `-Type`.